// deleghebowcl/src/app/model/cittadino.ts
/**
 * Deleghe
 * Api risorse per gestire le deleghe e le dichiarazioni (deleghe minori)
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/*******************************************************************************
* Copyright Regione Piemonte - 2022
* SPDX-License-Identifier: EUPL-1.2
******************************************************************************/
import {Documento} from './documento';


export class Cittadino {
  /**
   * codice fiscale
   */
  codice_fiscale?: string;
  /**
   * nome
   */
  nome?: string;
  /**
   * cognome
   */
  cognome?: string;
  /**
   * data di nascita
   */
  data_nascita?: string;
  /**
   * comune di nascita
   */
  comune_nascita?: string;
  /**
   * sesso questo campo non è obbligatorio
   */
  sesso?: string;

  asl?: string;
  /**
   * idAura questo campo non è obbligatorio
   */
  id_aura?: number;

  cit_id?: number;

  documento?: Documento;
}

export class CittadinoOrdered {
  citizen?: Cittadino;
  order?: number;
  nPratica?: number;
}