// deleghebowcl/src/app/model/documentoFile.ts
/**
 * Deleghe
 * Api risorse per gestire le deleghe e le dichiarazioni (deleghe minori)
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/*******************************************************************************
* Copyright Regione Piemonte - 2022
* SPDX-License-Identifier: EUPL-1.2
******************************************************************************/

import { Entita } from "./entita";


export interface DocumentoFile extends Entita{
    nome?: string;
    dimensione_in_bytes?: number;
    base_64_file?: string;
    //bytes?: FileList;
    bytes?: string;
    file_id?: number;
}